% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{sandboxing}
\alias{sandboxing}
\alias{eval.secure}
\title{Sandboxing}
\usage{
eval.secure(...)
}
\arguments{
\item{...}{arguments passed to \link[unix:eval_safe]{unix::eval_safe}}
}
\description{
This function has been superseded by the \link[unix:eval_safe]{unix::eval_safe} function.
}
