\name{setgid}
\alias{getgid}
\alias{setgid}
\title{Get/Set GID}
\usage{
  setgid(gid, verbose = FALSE)
}
\arguments{
  \item{gid}{group ID. Must be integer.}

  \item{verbose}{print some C output (TRUE/FALSE)}
}
\description{
  Wrappers for getgid and setgid in Linux.
}
\references{
  http://manpages.ubuntu.com/manpages/precise/man2/setgid.2.html
}

