\name{aa_find_mountpoint}
\alias{aa_find_mountpoint}
\title{Find the apparmor mountpoint}
\usage{
  aa_find_mountpoint(verbose = TRUE)
}
\arguments{
  \item{verbose}{print some C output (TRUE/FALSE)}
}
\value{
  location of mountpoint
}
\description{
  Find the apparmor mountpoint
}
\references{
  http://manpages.ubuntu.com/manpages/precise/man2/aa_find_mountpoint.2.html
}

