% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.R
\name{eval.secure}
\alias{eval.secure}
\title{eval.secure}
\description{
The \code{eval.secure} function has moved into the unix package and is now
an alias for \link[unix:eval_fork]{unix::eval_safe}. Please switch over
to this new function.
}
