% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{kill}
\alias{getgid}
\alias{getpgid}
\alias{getpriority}
\alias{getuid}
\alias{kill}
\alias{setgid}
\alias{setpgid}
\alias{setpriority}
\alias{setuid}
\title{Unix Process Utilities}
\usage{
kill(pid, signal = SIGTERM, verbose = FALSE)

setuid(uid, verbose = FALSE)

getuid()

setgid(gid, verbose = FALSE)

getgid()

setpgid(verbose = FALSE)

getpgid()

setpriority(prio, verbose = FALSE)

getpriority()
}
\arguments{
\item{pid}{process ID}

\item{signal}{kill signal}

\item{verbose}{emit some debugging output in C}

\item{uid}{user ID}

\item{gid}{group ID}

\item{prio}{priority value}
}
\description{
Read or set process properties.
}
\references{
Ubuntu Manpage: \code{kill} - \emph{send signal to a process}. \url{http://manpages.ubuntu.com/manpages/precise/man2/kill.2.html}.
}

