\name{setinteractive}
\alias{setinteractive}
\title{Toggle interactive mode in R.}
\usage{
  setinteractive(interactivity)
}
\arguments{
  \item{interactivity}{interactivity flag. TRUE or FALSE.}
}
\description{
  Toggle R interactive mode by setting C level global
  variable R_Interactive. See ?interactive for details
}

